requirements.txt
Flask==2.2.2
pandas==1.5.3
psycopg2-binary==2.9.5